requirements-dev.txt
# Development Dependencies for Face Recognition System
# These packages are used for development, testing, and code quality

# ==========================================
# Testing Framework
# ==========================================
pytest==7.4.3
pytest-cov==4.1.0
pytest-xdist==3.5.0  # Parallel test execution
pytest-mock==3.12.0
pytest-timeout==2.2.0
pytest-asyncio==0.21.1
pytest-watch==4.2.0  # Watch mode for tests
hypothesis==6.92.2  # Property-based testing
faker==20.1.0  # Generate fake data for tests

# ==========================================
# Code Quality & Linting
# ==========================================
black==23.12.1  # Code formatter
isort==5.13.2  # Import sorter
flake8==6.1.0  # Linting
flake8-docstrings==1.7.0
flake8-bugbear==23.12.2
flake8-comprehensions==3.14.0
flake8-simplify==0.21.0
pylint==3.0.3  # Advanced linting
autopep8==2.0.4  # Auto-formatting

# ==========================================
# Type Checking
# ==========================================
mypy==1.7.1
types-requests==2.31.0.10
types-Pillow==10.1.0.2
types-setuptools==69.0.0.0

# ==========================================
# Security Scanning
# ==========================================
bandit==1.7.5  # Security linter
safety==2.3.5  # Dependency vulnerability scanner
pip-audit==2.6.1  # Audit Python packages

# ==========================================
# Documentation
# ==========================================
sphinx==7.2.6
sphinx-rtd-theme==2.0.0
sphinx-autodoc-typehints==1.25.2
mkdocs==1.5.3
mkdocs-material==9.5.3
pdoc3==0.10.0

# ==========================================
# Code Analysis
# ==========================================
radon==6.0.1  # Code complexity metrics
vulture==2.10  # Find dead code
pydocstyle==6.3.0  # Docstring style checker
interrogate==1.5.0  # Check docstring coverage

# ==========================================
# Pre-commit Hooks
# ==========================================
pre-commit==3.6.0

# ==========================================
# Profiling & Performance
# ==========================================
py-spy==0.3.14  # Sampling profiler
memory-profiler==0.61.0
line-profiler==4.1.1
pyinstrument==4.6.1

# ==========================================
# Debugging
# ==========================================
ipdb==0.13.13  # IPython debugger
ipython==8.18.1  # Enhanced Python shell
pdbpp==0.10.3  # Enhanced pdb

# ==========================================
# Development Utilities
# ==========================================
python-dotenv==1.0.0  # Environment variables
watchdog==3.0.0  # File system events
colorama==0.4.6  # Colored terminal output
tqdm==4.66.1  # Progress bars
click==8.1.7  # CLI framework
rich==13.7.0  # Rich terminal output

# ==========================================
# API Testing
# ==========================================
httpx==0.25.2  # Async HTTP client
responses==0.24.1  # Mock HTTP responses
pytest-httpx==0.27.0

# ==========================================
# Mocking & Fixtures
# ==========================================
factory-boy==3.3.0  # Test fixtures
freezegun==1.4.0  # Time mocking
responses==0.24.1  # HTTP mocking

# ==========================================
# Coverage & Reporting
# ==========================================
coverage[toml]==7.3.3
pytest-html==4.1.1  # HTML test reports
pytest-json-report==1.5.0  # JSON test reports

# ==========================================
# Build & Packaging
# ==========================================
build==1.0.3
twine==4.0.2
setuptools==69.0.2
wheel==0.42.0

# ==========================================
# Git Hooks & Automation
# ==========================================
commitizen==3.13.0  # Commit message standard
bump2version==1.0.1  # Version bumping
gitpython==3.1.40

# ==========================================
# Notebook Development
# ==========================================
jupyter==1.0.0
jupyterlab==4.0.9
notebook==7.0.6
nbconvert==7.12.0
nbformat==5.9.2

# ==========================================
# Data Validation & Schema
# ==========================================
pydantic[email]==2.5.3
marshmallow==3.20.1
cerberus==1.3.5

# ==========================================
# Additional Testing Tools
# ==========================================
locust==2.19.1  # Load testing
tavern==2.8.0  # API testing
schemathesis==3.27.1  # API schema testing
aiohttp==3.9.1 # Async HTTP client/server
async-timeout==4.0.3
